14. Zrobiono z klasy Node wewnętrzną klasę klasy IntLinkedList
